+++ lib/src/friendly_to_string.dart
@@ -1,0 +1,6 @@
+abstract class FriendlyToString {
+  toJson();
+
+  @override
+  String toString() => '${this.runtimeType}${this.toJson()}';
+}