+++ lib/src/meta.dart
@@ -1,10 +1,11 @@
+import 'package:json_api_document/src/friendly_to_string.dart';
 import 'package:json_api_document/src/naming.dart';
 import 'package:json_api_document/src/readonly_map.dart';
 
 /// A Meta information object.
 ///
 /// http://jsonapi.org/format/#document-meta
-class Meta extends ReadonlyMap<String, dynamic> {
+class Meta extends ReadonlyMap<String, dynamic> with FriendlyToString {
   Meta(Map<String, dynamic> meta) : super(meta) {
     if (isEmpty) throw ArgumentError('Empty meta');
     keys.forEach(Naming().enforce);