+++ lib/src/relationships.dart
@@ -1,8 +1,10 @@
+import 'package:json_api_document/src/friendly_to_string.dart';
 import 'package:json_api_document/src/naming.dart';
 import 'package:json_api_document/src/readonly_map.dart';
 import 'package:json_api_document/src/relationship.dart';
 
-class Relationships extends ReadonlyMap<String, Relationship> {
+class Relationships extends ReadonlyMap<String, Relationship>
+    with FriendlyToString {
   Relationships(Map<String, Relationship> relationships)
       : super(relationships) {
     keys.forEach(Naming().enforce);